requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
fpdf==1.7.2
html5lib==1.0.1
idna==2.8
jieba3k==0.35.1
lxml==4.4.1
newspaper3k==0.2.8
nltk==3.4.5
Pillow==6.1.0
pkg-resources==0.0.0
pyPdf==1.13
PyQt5==5.13.0
PyQt5-sip==4.19.18
python-dateutil==2.8.0
PyYAML==5.1.2
reportlab==3.5.23
requests==2.22.0
requests-file==1.4.3
six==1.12.0
soupsieve==1.9.3
text2pdf==0.2
textblob==0.15.3
textile==3.0.4
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.25.3
webencodings==0.5.1
